Chrome extension to download Google Group contents
I've found clicking the "Next" button via JS to be unreliable for
Google Groups so I'm keeping it simple and just extracting
conversation URLs from the current index page. Hopefully we'll
come up with a more reliable method in a future iteration.

The icon is based on https://github.com/apancik/public-domain-icons.