Resolve issues with downloading the wrong URL
With the previous version, we were seeing issues with the
downloaded content not reflecting the intended URL, probably
due to the update to the tab URL not having been complete before
the start of the download.

We resolved this by taking advantage of callbacks/listeners allowing
us to check on the status of the tab. (Not sure if the new Promise
based API offers this; we wrapped the old callback-based API in a
Promise.) Thank you to a StackOverflow user for the insight on how
to handle this.

Also adding more logging to help with understanding issues.